Trying casting to Integer to remove id

+++ src/org/kihara/util/TicketManager.java
@@ -126,7 +126,7 @@ public class TicketManager<T> {
             fileHashMap.remove(id);
         }
         else if (hasTicket(id)) {
-            idSet.remove(id);
+            idSet.remove((Integer) id);
         }
     }
 }
